requirements.txt
streamlit==1.28.1
streamlit-chat==0.1.1
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2
groq==0.13.3
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
